+++ Sources/RetentionFixturesCrossModule/ExternalAssociatedType.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol ExternalAssociatedType {
+    associatedtype Value
+}