+++ Tests/RetentionFixtures/testRetainsExternalAssociatedTypeTypeAlias.swift
@@ -1,0 +1,12 @@
+import Foundation
+import RetentionFixturesCrossModule
+
+struct Fixture110: ExternalAssociatedType {
+    typealias Value = Void
+}
+
+public struct Fixture110Retainer {
+    public func someFunc() {
+        print(Fixture110.self)
+    }
+}